# docs/trading-openapi.yaml
openapi: 3.1.0
info:
  title: Yogimangchi Trading API
  version: v1
servers:
- url: /
paths:
  /api/v1/trading/accounts/guest:
    post:
      tags:
      - trading-account-controller
      summary: Create a guest simulation account with 10000 USDT
      description: Returns a 7-day opaque bearer credential once. Losing it means
        losing access; never share or log it.
      operationId: createGuest
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestAccountResponse"
  /api/v1/trading/account/positions/{positionId}/close:
    post:
      tags:
      - order-controller
      summary: Close all or part of an owned position at the fresh server mark price
      description: "Omit body/quantity for the entire remaining lot. Quantity is positive\
        \ with at most 8 decimal places; no minimum exit notional. Only unreserved\
        \ quantity may close; excess returns 409 CLOSE_QUANTITY_EXCEEDED. Cancel pending\
        \ CLOSE orders before closing their reserved quantity. LONG closes by selling,\
        \ SHORT by buying. All held symbols must be fresh. Same key with a different\
        \ quantity returns 409; 404 for another account's position."
      operationId: close
      parameters:
      - name: positionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosePositionRequest"
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security:
      - guestBearer: []
  /api/v1/trading/account/positions/{positionId}/close-orders:
    post:
      tags:
      - order-controller
      summary: Reserve a LIMIT CLOSE for part or all of an owned position
      description: "Returns PENDING; creates no new position and reserves quantity,\
        \ not wallet margin. Side/leverage/symbol are inherited from the target lot.\
        \ LONG CLOSE triggers at Mark >= limit; SHORT CLOSE at Mark <= limit. The\
        \ triggering Mark is the execution price. Omit quantity for the entire remaining\
        \ lot. Existing reservations cannot be exceeded (409). Fresh portfolio prices\
        \ and READY engine required (503 otherwise). Same account/key/payload replays\
        \ the current order; different payload returns 409. Other account's target\
        \ returns 404."
      operationId: limitClose
      parameters:
      - name: positionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitCloseRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security:
      - guestBearer: []
  /api/v1/trading/account/orders:
    get:
      tags:
      - order-controller
      summary: "Read own order/fill history, newest first"
      description: "Use the last returned orderId as beforeId for the next page; limit\
        \ 1-100, default 50."
      operationId: history
      parameters:
      - name: beforeId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
      - guestBearer: []
    post:
      tags:
      - order-controller
      summary: OPEN a new lot with MARKET or reserve a PENDING LIMIT OPEN order
      description: "Idempotency-Key: 8-100 letters, digits, underscore or hyphen.\
        \ Replay returns the same order with its current state. 400 invalid request;\
        \ 401 credential; 409 margin/risk/idempotency conflict; 503 unavailable price\
        \ or recovering engine. No client execution price."
      operationId: create
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security:
      - guestBearer: []
  /api/v1/trading/account/orders/{orderId}/cancel:
    post:
      tags:
      - order-controller
      summary: Cancel an owned pending order and release its margin or close quantity
      description: Repeated cancel is idempotent. A fill that wins the race returns
        409. Cancellation remains available during price/Redis outages.
      operationId: cancel
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security:
      - guestBearer: []
  /api/v1/trading/status:
    get:
      tags:
      - trading-engine-status-controller
      summary: Read trigger infrastructure status
      description: "READY means the worker is running with healthy Redis. Each order\
        \ still requires fresh prices and a processed price vector; RECOVERING blocks\
        \ new exposure and manual closes, but not cancellations."
      operationId: status
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /api/v1/trading/account/wallet:
    get:
      tags:
      - trading-account-controller
      summary: Read the authenticated guest's USDT wallet and current valuation
      description: "401: missing/expired credential; 409 TRADING_BUSY: retry after\
        \ lock contention; 503: authentication storage unavailable."
      operationId: wallet
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
      security:
      - guestBearer: []
  /api/v1/trading/account/summary:
    get:
      tags:
      - trading-account-controller
      summary: "Read a coherent account, wallet and open-position snapshot"
      description: "401: missing/expired credential; 409 TRADING_BUSY: retry after\
        \ lock contention; 503: authentication storage unavailable."
      operationId: summary
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSummaryResponse"
      security:
      - guestBearer: []
  /api/v1/trading/account/positions:
    get:
      tags:
      - trading-account-controller
      summary: Read the authenticated guest's open position lots
      description: "401: missing/expired credential; 409 TRADING_BUSY: retry after\
        \ lock contention; 503: authentication storage unavailable."
      operationId: positions
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PositionResponse"
      security:
      - guestBearer: []
  /api/v1/trading/account/orders/pending:
    get:
      tags:
      - order-controller
      summary: Read own pending OPEN/CLOSE orders and their margin/quantity reservations
      operationId: pending
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
      - guestBearer: []
  /api/v1/symbols:
    get:
      tags:
      - trading-symbol-controller
      summary: 거래 가능한 종목 목록
      description: 인증 없이 ACTIVE 종목을 Symbol 오름차순으로 조회합니다. 종목이 없으면 빈 배열을 반환합니다.
      operationId: findActiveSymbols
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingSymbolResponse"
components:
  schemas:
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type:
          - string
          - "null"
        status:
          type: integer
          format: int32
        detail:
          type:
          - string
          - "null"
        instance:
          type:
          - string
          - "null"
          format: uri
        properties:
          type:
          - object
          - "null"
          additionalProperties: {}
    GuestAccountResponse:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        status:
          type: string
        quoteAsset:
          type: string
        initialBalance:
          type: string
        accessToken:
          type: string
          description: "256-bit opaque guest credential. Returned once; send as Authorization:\
            \ Bearer <token>."
        tokenType:
          type: string
        expiresAt:
          type: string
          format: date-time
    ClosePositionRequest:
      type: object
      properties:
        quantity:
          type: string
          description: "Quantity to close, maximum 8 decimal places. Omit for the\
            \ entire remaining lot."
          example: "0.01"
    FillResponse:
      type: object
      properties:
        fillId:
          type: integer
          format: int64
        quantity:
          type: string
        price:
          type: string
        priceEventTime:
          type: string
          format: date-time
        executedAt:
          type: string
          format: date-time
        realizedPnl:
          type: string
    OrderResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        positionId:
          type: integer
          format: int64
        tradingSymbolId:
          type: integer
          format: int64
        type:
          type: string
        action:
          type: string
        side:
          type: string
        status:
          type: string
        quantity:
          type: string
        leverage:
          type: integer
          format: int32
        limitPrice:
          type: string
        reservedMargin:
          type: string
        reservedCloseQuantity:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        reason:
          type: string
        fill:
          $ref: "#/components/schemas/FillResponse"
    LimitCloseRequest:
      type: object
      properties:
        quantity:
          type: string
          description: "Positive domain quantity, maximum 8 decimals. Omit to reserve\
            \ the entire remaining lot; existing reservations may cause 409."
          example: "0.01"
        limitPrice:
          type: string
          description: "Required positive USDT limit, maximum 18 decimals. LONG CLOSE\
            \ triggers at or above; SHORT CLOSE at or below. Execution uses the triggering\
            \ server Mark."
          example: "70000"
    CreateOrderRequest:
      type: object
      description: OPEN a new independent position lot. Reduce existing exposure through
        the position close or close-orders endpoint.
      properties:
        type:
          type: string
          enum:
          - MARKET
          - LIMIT
        tradingSymbolId:
          type: integer
          format: int64
        side:
          type: string
          enum:
          - LONG
          - SHORT
        quantity:
          type: string
          description: "Domain asset quantity, maximum 8 decimal places"
          example: "0.001"
        leverage:
          type: integer
          format: int32
          description: Lot leverage; fixed after opening
          maximum: 20
          minimum: 1
        limitPrice:
          type: string
          description: Required only for LIMIT. LONG fills at or below; SHORT at or
            above this price
          example: "60000"
    StatusResponse:
      type: object
      properties:
        status:
          type: string
    WalletResponse:
      type: object
      description: USDT amounts are decimal strings. Valuation-dependent fields are
        null when any open position price is unavailable.
      properties:
        quoteAsset:
          type: string
        balance:
          type: string
        usedMargin:
          type: string
        reservedMargin:
          type: string
        realizedPnl:
          type: string
        unrealizedPnl:
          type: string
        equity:
          type: string
        availableBalance:
          type: string
        valuationStatus:
          type: string
    AccountSummaryResponse:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        wallet:
          $ref: "#/components/schemas/WalletResponse"
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionResponse"
        valuedAt:
          type: string
          format: date-time
    PositionResponse:
      type: object
      description: Independent position lot. quantity/margin are remaining values;
        realizedPnl is cumulative. markPrice/unrealizedPnl are null unless the server
        price is FRESH.
      properties:
        id:
          type: integer
          format: int64
        tradingSymbolId:
          type: integer
          format: int64
        side:
          type: string
        quantity:
          type: string
        reservedCloseQuantity:
          type: string
          description: "Quantity reserved by PENDING CLOSE orders, independent of\
            \ wallet reservedMargin"
        freeCloseQuantity:
          type: string
          description: Remaining quantity minus reservedCloseQuantity; available for
            MARKET or new LIMIT CLOSE
        entryPrice:
          type: string
        leverage:
          type: integer
          format: int32
        margin:
          type: string
        status:
          type: string
        realizedPnl:
          type: string
        markPrice:
          type: string
        unrealizedPnl:
          type: string
        priceStatus:
          type: string
        priceEventTime:
          type: string
          format: date-time
        openedAt:
          type: string
          format: date-time
    TradingSymbolResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Provider Symbol 변경과 무관하게 유지되는 내부 종목 ID
          example: 1
        symbol:
          type: string
          example: BTC
        name:
          type: string
          example: Bitcoin
        quoteAsset:
          type: string
          example: USDT
        displaySymbol:
          type: string
          description: 화면 표시용 거래쌍
          example: BTC / USDT
  securitySchemes:
    guestBearer:
      type: http
      description: Opaque Guest Trading MVP credential; not a Content JWT. Issued
        by POST /api/v1/trading/accounts/guest.
      scheme: bearer